extract/requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
httpx==0.23.3
boto3==1.26.41
s3fs==0.4.2
fastapi-utils==0.2.1
requests==2.28.2
kafka-python==2.0.1